+++ src/main/scala/com/github/vitalsoftware/scalaredox/models/HasVisitInfo.scala
@@ -1,5 +1,9 @@
 package com.github.vitalsoftware.scalaredox.models
 
+trait HasVisit {
+  def Visit: Option[VisitLike]
+}
+
 trait HasVisitInfo {
   def Visit: Option[VisitInfo]
 }
