+++ src/main/scala/com/github/vitalsoftware/scalaredox/models/Order.scala
@@ -142,35 +142,3 @@ object OrderingFacility extends RobustPrimitives
 )
 
 object Order extends RobustPrimitives
-
-trait OrdersMessageLike extends MetaLike with HasPatient with HasVisitInfo {
-  def Meta: Meta
-  def Patient: Patient
-  def Visit: Option[VisitInfo]
-  def Orders: Seq[Order]
-}
-
-/**
- * Order messages communicate details of diagnostic tests such as labs, radiology imaging, etc.
- */
-@jsonDefaults case class OrderMessage(
-  Meta: Meta,
-  Patient: Patient,
-  Visit: Option[VisitInfo] = None,
-  Order: Order
-) extends MetaLike
-    with OrdersMessageLike {
-  def Orders = Seq(Order)
-}
-
-object OrderMessage extends RobustPrimitives
-
-@jsonDefaults case class GroupedOrdersMessage(
-  Meta: Meta,
-  Patient: Patient,
-  Visit: Option[VisitInfo] = None,
-  Orders: Seq[Order] = Seq.empty
-) extends MetaLike
-    with OrdersMessageLike
-
-object GroupedOrdersMessage extends RobustPrimitives
